+++ src/mapimage/mapImageUpdater.py
@@ -1,3 +1,4 @@
+from appContainer import component
 from config.config import Config
 from mapimage.mapImageGenerator import MapImageGenerator
 from world.tickCounter import TickCounter
@@ -5,6 +6,7 @@ from world.tickCounter import TickCounter
 
 # @author Daniel McCoy Stephenson
 # @since February 2nd, 2023
+@component
 class MapImageUpdater:
     def __init__(self, tickCounter: TickCounter, config: Config):
         self.tickCounter = tickCounter